requirements.txt
certifi==2018.11.29
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
rpi.gpio==0.6.5